Fix for GCD issue Fix for routes disappearing forever
git-svn-id: http://pocketgnome.googlecode.com/svn/trunk/pocketgnome@202 4a9c585c-cb2d-11dd-a3c7-fda619b15ae3

+++ PlayerDataController.m
@@ -568,7 +568,7 @@ static PlayerDataController* sharedController = nil;
     [[controller wowMemoryAccess] saveDataForAddress: ([self baselineAddress] + BaseField_Facing_Horizontal) Buffer: (Byte*)&direction BufLength: sizeof(direction)];
 }
 
-- (void)setMovementFlags:(UInt32)movementFlags {
+- (void)setMovementFlags:(UInt8)movementFlags {
     [[controller wowMemoryAccess] saveDataForAddress: ([self baselineAddress] + BaseField_MovementFlags) Buffer: (Byte*)&movementFlags BufLength: sizeof(movementFlags)];
 }
